+++ src/CustomizeRecipe/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.3.0] U44-535842-S
+- fixed buildings so they don't force store outputs (effects GlassForge and UraniumCentrifuge)
+- recipes with missing storeElement property will default to true, if the element is liquid
+- added new setting HEPout (only for buildings that can produce them)
+
 ## [1.0.2.0] U41-498381-S
 - added sanity checks